Refactor the codebase to add custom hooks and clean the components and pages files
