+++ components/user/ChatInput.tsx
@@ -1,11 +1,7 @@
 "use client";
 
-import { chatInputSchema } from "@/lib/validations/auth";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { useSession } from "next-auth/react";
-import React, { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
+import React from "react";
 import {
   Form,
   FormControl,
@@ -15,62 +11,14 @@ import {
 } from "../ui/form";
 import { Button } from "../ui/Button";
 import { Input } from "../ui/Input";
-import {
-  collection,
-  doc,
-  onSnapshot,
-  serverTimestamp,
-  setDoc,
-} from "firebase/firestore";
-import { db } from "@/firebase";
-import { userRef } from "@/lib/converters/User";
-
-type CahtInput = z.infer<typeof chatInputSchema>;
-interface InputUser {
-  name: string;
-  email: string;
-  image: string;
-}
+import { useUserData } from "@/hooks/useUserData";
+import { useSubmitChatMessage } from "@/hooks/useSubmitChatMessage";
 
 const ChatInput = ({ chatId }: { chatId: string }) => {
   const { data: session } = useSession();
-  const [inputUser, setInputUser] = useState<InputUser>();
 
-  useEffect(() => {
-    if (session?.user?.id) {
-      const inputUserRef = userRef(session.user.id);
-      const unsubscribe = onSnapshot(inputUserRef, (doc) => {
-        setInputUser(doc.data());
-      });
-      return () => unsubscribe();
-    }
-  }, [session?.user?.id]);
-
-  const form = useForm<CahtInput>({
-    resolver: zodResolver(chatInputSchema),
-    defaultValues: {
-      input: "",
-    },
-  });
-
-  async function onSubmit(data: CahtInput) {
-    const inputCopy = data.input.trim();
-    form.reset();
-
-    if (!inputCopy || !session?.user?.id || !chatId) return;
-    const messageRef = doc(collection(db, "chats", chatId, "messages"));
-    await setDoc(messageRef, {
-      text: inputCopy,
-      isRead: false,
-      timestamp: serverTimestamp(),
-      user: {
-        name: inputUser?.name,
-        email: session.user.email,
-        id: session.user.id,
-        image: inputUser?.image,
-      },
-    });
-  }
+  const inputUser = useUserData();
+  const { form, onSubmit } = useSubmitChatMessage(chatId, session, inputUser);
 
   return (
     <div className="sticky bottom-0">
